backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==37.0.2
cycler==0.11.0
dnspython==2.2.1
ecdsa==0.17.0
fastapi==0.75.2
fonttools==4.33.3
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
mongoengine==0.24.1
motor==3.0.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
passlib==1.7.4
Pillow==9.1.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pykerberos==1.2.4
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
python-snappy==0.6.1
pytz==2022.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tqdm==4.64.0
typing_extensions==4.2.0
uvicorn==0.17.6